fix(mu): remove logs and cu url overwrites
