+++ servers/mu/src/domain/clients/scheduler.js
@@ -1,49 +1,7 @@
 import { identity } from 'ramda'
 import { of, fromPromise, Rejected } from 'hyper-async'
+import { backoff, okRes } from '../utils'
 
-const okRes = (res) => {
-  if (res.ok) return res
-  throw res
-}
-export const backoff = (
-  fn,
-  { maxRetries = 3, delay = 500, log, name }
-) => {
-  /**
-   * Recursive function that recurses with exponential backoff
-   */
-  console.log('BACKOFF')
-  const action = (retry, delay) => {
-    return Promise.resolve()
-      .then(fn)
-      .catch((err) => {
-        console.log('BACKOFF ERROR')
-        // Reached max number of retries
-        if (retry >= maxRetries) {
-          log(`(${name}) Reached max number of retries: ${maxRetries}. Bubbling err`)
-          return Promise.reject(err)
-        }
-
-        const newRetry = retry + 1
-        const newDelay = delay + delay
-        log(`(${name}) Backing off -- retry ${newRetry} starting in ${newDelay} milliseconds...`)
-        return new Promise((resolve, reject) =>
-          setTimeout(
-          /**
-           * increment the retry count Retry with an exponential backoff
-           */
-            () => action(newRetry, newDelay).then(resolve).catch(reject),
-            /**
-             * Retry in {delay} milliseconds
-             */
-            delay
-          )
-        )
-      })
-  }
-
-  return action(0, delay)
-}
 function writeDataItemWith ({ fetch, logger }) {
   return async ({ data, suUrl }) => {
     return of(Buffer.from(data, 'base64'))
@@ -111,7 +69,6 @@ function writeAssignmentWith ({ fetch, logger }) {
           url += `&exclude=${exclude}`
         }
 
-        console.log({ url })
         return backoff(
           () => fetch(url, {
             method: 'POST',
@@ -121,12 +78,10 @@ function writeAssignmentWith ({ fetch, logger }) {
             },
             redirect: 'manual'
           }).then((res) => {
-            console.log('BACKOFF RES', { res })
             return okRes(res)
           }),
           { maxRetries: 5, delay: 500, log: logger, name: `forwardAssignment(${JSON.stringify({ suUrl, processId, txId })})` }
         ).then(response => {
-          console.log({ response })
           if ([307, 308].includes(response.status)) {
             const newUrl = response.headers.get('Location')
             return fetch(newUrl, {
@@ -148,7 +103,6 @@ function writeAssignmentWith ({ fetch, logger }) {
       .bichain(
         (err) => Rejected(JSON.stringify(err)),
         fromPromise(async (res) => {
-          console.log({ suRes: res })
           if (!res?.ok) {
             const text = await res.text()
             throw new Error(`${res.status}: ${text}`)
@@ -163,21 +117,17 @@ function writeAssignmentWith ({ fetch, logger }) {
 
 function fetchSchedulerProcessWith ({ fetch, logger, setByProcess, getByProcess }) {
   return (processId, suUrl) => {
-    console.log('abc2', { processId, suUrl, fetch: fetch.toString() })
     return getByProcess(processId)
       .then(cached => {
-        console.log('abc3', { cached })
         if (cached) {
           logger(`cached process found ${processId}`)
           return cached
         }
 
         logger(`${suUrl}/processes/${processId}`)
-        // suUrl = 'http://localhost:9000'
         return fetch(`${suUrl}/processes/${processId}`)
           .then(res => res.json())
           .then(res => {
-            console.log('abc4', { res })
             if (res) {
               return setByProcess(processId, res).then(() => res)
             }
